% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_manipulation.R
\name{code_missing}
\alias{code_missing}
\title{Fix Missing Values inside of a Vector}
\usage{
code_missing(x, value = -1L)
}
\arguments{
\item{x}{A vector}

\item{value}{The value that denotes a missing value
inside of the data.}
}
\value{
A vector with missing values appropriately
typed.
}
\description{
Within this function, we will modify
the initial input vector by searching for
what the researcher specified as the missing 
value and instead subsitute it for R's
missing data type.
}
\examples{
x = c(1,2, 3, -1)

code_missing(x) 

code_missing(x, value = 2)
}
